- Добавлен отдельный файл для фикстур - Удалены фикстуры из файлов папки tests
