+++ pyproject.toml
@@ -8,6 +8,7 @@ authors = [
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
+    "pytest (>=8.4.1,<9.0.0)"
 ]
